clean up handler script and move to libraries folder
